Move event classes to Event submodule. Update header paths.
