+++ include/SimCore/DetectorConstruction.h
@@ -25,7 +25,7 @@
 /*~~~~~~~~~~~~~~~*/
 /*   Framework   */
 /*~~~~~~~~~~~~~~~*/
-#include "Framework/Parameters.h" 
+#include "Framework/Configure/Parameters.h" 
 
 namespace ldmx {
