+++ include/SimCore/MultiParticleGunPrimaryGenerator.h
@@ -39,7 +39,7 @@
 //-------------//
 #include "SimCore/UserPrimaryParticleInformation.h"
 #include "SimCore/PrimaryGenerator.h"
-#include "Framework/Parameters.h"
+#include "Framework/Configure/Parameters.h"
 
 namespace ldmx {
