+++ include/SimCore/UserAction.h
@@ -18,7 +18,7 @@
 /*~~~~~~~~~~~~~~~*/
 /*   Framework   */
 /*~~~~~~~~~~~~~~~*/
-#include "Framework/Parameters.h" 
+#include "Framework/Configure/Parameters.h" 
 
 // Forward Declarations
 class G4Event;
