+++ src/G4eDarkBremsstrahlungModel.cxx
@@ -7,7 +7,7 @@
 
 #include "SimCore/G4eDarkBremsstrahlungModel.h"
 #include "SimCore/G4APrime.h"
-#include "Framework/Exception.h"
+#include "Framework/Exception/Exception.h"
 
 // Geant4
 #include "G4PhysicalConstants.hh"
