+++ src/MagneticFieldMap3D.cxx
@@ -1,5 +1,5 @@
 #include "SimCore/MagneticFieldMap3D.h"
-#include "Framework/Exception.h"
+#include "Framework/Exception/Exception.h"
 
 // STL
 #include <fstream>
