+++ src/RootCompleteReSim.cxx
@@ -21,7 +21,7 @@
 //   ldmx-sw   //
 //-------------//
 #include "SimCore/SimParticle.h"
-#include "Framework/Parameters.h"
+#include "Framework/Configure/Parameters.h"
 
 namespace ldmx {
